+++ CHANGELOG.md
@@ -5,7 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Unreleased
---
+- Implemented Kingdom interface
+- Removed App interface
+- Removed Veneer Context
+- Removed appName from Environment config
+- Simplified Genesis Hub initialisation
+- Moved HTTP Kernel to Harvest as Kingdom Runtime
 
 ---